feat: hook useResetParams
