+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useAquariumDetails } from './useAquariumDetails';
+export { useResetParams } from './useResetParams';
